requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
django-filter
python-decouple
twilio
drf-yasg